Commonize Mobile and Desktop TokenControls
