+++ packages/frontend/src/components/chart/MobileTokenControls.tsx
@@ -3,32 +3,15 @@ import React from 'react'
 import { HorizontalSeparator } from '../HorizontalSeparator'
 import { ExpandIcon } from '../icons/Expand'
 import { SlideCard } from '../SlideCard'
-import { SelectedTokenButton, TokenControl } from './DesktopTokenControls'
+import {
+  getParts,
+  SelectedTokenButton,
+  TokenCell,
+  TokenControl,
+} from './CommonTokenControls'
 
 export interface MobileTokenControlsProps {
   tokens?: TokenControl[]
-}
-
-function TokenCell({ token }: { token: TokenControl }) {
-  return (
-    <label className="flex cursor-pointer select-none items-center gap-1.5">
-      <input
-        className="peer hidden"
-        name="token"
-        type={'radio'}
-        autoComplete="off"
-        value={token.symbol}
-        data-tvl-endpoint={token.tvlEndpoint}
-      />
-      <img
-        src={`https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/${token.address}/logo.png`}
-        className="h-4 w-4 rounded-full"
-      />
-      <p className="text-sm" key={token.symbol}>
-        <span className={'font-bold'}>{token.name}</span> ({token.symbol})
-      </p>
-    </label>
-  )
 }
 
 export function MobileTokenControls({ tokens }: MobileTokenControlsProps) {
@@ -36,23 +19,7 @@ export function MobileTokenControls({ tokens }: MobileTokenControlsProps) {
     return null
   }
 
-  const parts = [
-    {
-      title: 'Natively Minted Tokens',
-      titleColor: 'text-[#FF6DCD]',
-      tokens: tokens.slice(0, 5),
-    },
-    {
-      title: 'Externally Bridged Tokens',
-      titleColor: 'text-yellow-200',
-      tokens: tokens.slice(15, 25),
-    },
-    {
-      title: 'Canonically Bridged Tokens',
-      titleColor: 'text-[#D98EFF]',
-      tokens: tokens.slice(30, 45),
-    },
-  ]
+  const parts = getParts(tokens)
 
   return (
     <div className="md:hidden">
